extract `FeedStoreSpy` into its own file to avoid duplication
